Initialising the artemis notification plugin
