+++ .gitignore
@@ -1,35 +1,16 @@
-# Compiled class file
-*.class
+.gradle
+/build/
 
-# Log file
-*.log
+# Ignore Gradle GUI config
+gradle-app.setting
 
-# BlueJ files
-*.ctxt
+# Avoid ignoring Gradle wrapper jar file (.jar files are usually ignored)
+!gradle-wrapper.jar
 
-# Mobile Tools for Java (J2ME)
-.mtj.tmp/
+# Cache of project
+.gradletasknamecache
 
-# Package Files #
-*.jar
-*.war
-*.nar
-*.ear
-*.zip
-*.tar.gz
-*.rar
+# # Work around https://youtrack.jetbrains.com/issue/IDEA-116898
+gradle/wrapper/gradle-wrapper.properties
 
-# virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
-hs_err_pid*
-
-# IDE files
-*.iml
-\.idea
-
-# Build files
-target
-work
-backup
-.settings/
-*.project
-*.classpath
+.idea
