+++ gradle.properties
@@ -1,0 +1,2 @@
+junit_version=5.9.1
+log4j_version=2.19.0
