+++ src/main/java/de/tum/in/ase/notification/exception/TestParsingException.java
@@ -1,0 +1,8 @@
+package de.tum.in.ase.notification.exception;
+
+public class TestParsingException extends RuntimeException {
+
+    public TestParsingException(Throwable cause) {
+        super(cause);
+    }
+}
