+++ src/main/java/de/tum/in/ase/notification/model/Commit.java
@@ -1,0 +1,34 @@
+package de.tum.in.ase.notification.model;
+
+public class Commit {
+
+    private String hash;
+
+    private String repositorySlug;
+
+    private String branchName;
+
+    public String getHash() {
+        return hash;
+    }
+
+    public void setHash(String hash) {
+        this.hash = hash;
+    }
+
+    public String getRepositorySlug() {
+        return repositorySlug;
+    }
+
+    public void setRepositorySlug(String repositorySlug) {
+        this.repositorySlug = repositorySlug;
+    }
+
+    public String getBranchName() {
+        return branchName;
+    }
+
+    public void setBranchName(String branchName) {
+        this.branchName = branchName;
+    }
+}
