--- src/main/java/de/tum/in/www1/jenkins/notifications/exception/TestParsingException.java
@@ -1,8 +1,0 @@
-package de.tum.in.www1.jenkins.notifications.exception;
-
-public class TestParsingException extends RuntimeException {
-
-    public TestParsingException(Throwable cause) {
-        super(cause);
-    }
-}
