--- src/main/java/de/tum/in/www1/jenkins/notifications/model/Commit.java
@@ -1,41 +1,0 @@
-package de.tum.in.www1.jenkins.notifications.model;
-
-import org.kohsuke.stapler.export.Exported;
-import org.kohsuke.stapler.export.ExportedBean;
-
-@ExportedBean(defaultVisibility = 2)
-public class Commit {
-
-    private String hash;
-
-    private String repositorySlug;
-
-    private String branchName;
-
-    @Exported
-    public String getHash() {
-        return hash;
-    }
-
-    public void setHash(String hash) {
-        this.hash = hash;
-    }
-
-    @Exported
-    public String getRepositorySlug() {
-        return repositorySlug;
-    }
-
-    public void setRepositorySlug(String repositorySlug) {
-        this.repositorySlug = repositorySlug;
-    }
-
-    @Exported
-    public String getBranchName() {
-        return branchName;
-    }
-
-    public void setBranchName(String branchName) {
-        this.branchName = branchName;
-    }
-}
