--- src/main/java/de/tum/in/www1/jenkins/notifications/model/CustomFeedback.java
@@ -1,34 +1,0 @@
-package de.tum.in.www1.jenkins.notifications.model;
-
-public class CustomFeedback {
-
-    private String name;
-
-    private boolean successful;
-
-    private String message;
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public boolean isSuccessful() {
-        return successful;
-    }
-
-    public void setSuccessful(boolean successful) {
-        this.successful = successful;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    public void setMessage(String message) {
-        this.message = message;
-    }
-}
