--- src/main/resources/META-INF/hudson.remoting.ClassFilter
@@ -1,3 +1,0 @@
-de.tum.in.ase.parser.domain.Report
-de.tum.in.ase.parser.domain.Issue
-com.google.gson.JsonArray
