--- src/main/resources/de/tum/in/www1/jenkins/notifications/Messages.properties
@@ -1,2 +1,0 @@
-SendTestResultsNotificationPostBuildTask.errors.postFailed=Sending test results failed (%d) with response: %s
-SendTestResultsNotificationPostBuildTask.DescriptorImpl.DisyplayName=Send Test Results Notification
