--- src/main/resources/de/tum/in/www1/jenkins/notifications/Messages_de.properties
@@ -1,2 +1,0 @@
-SendTestResultsNotificationPostBuildTask.errors.postFailed=Senden der Test Ergebnisse fehlgeschlagen (%d). Antwort war: %s
-SendTestResultsNotificationPostBuildTask.DescriptorImpl.DisyplayName=Test Ergebnis Benachrichtigung Senden
