--- src/main/resources/de/tum/in/www1/jenkins/notifications/SendTestResultsNotificationPostBuildTask/config.jelly
@@ -1,9 +1,0 @@
-<?jelly escape-by-default='true'?>
-<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:c="/lib/credentials" xmlns:d="jelly:define" xmlns:l="/lib/layout" xmlns:t="/lib/hudson" xmlns:f="/lib/form">
-    <f:entry title="${%Target}" field="notificationUrl">
-        <f:textbox />
-    </f:entry>
-    <f:entry title="${%Secret}" field="credentialsId">
-        <c:select/>
-    </f:entry>
-</j:jelly>
