--- src/main/resources/de/tum/in/www1/jenkins/notifications/SendTestResultsNotificationPostBuildTask/config.properties
@@ -1,2 +1,0 @@
-Target=Notification Target
-Secret=Token
