--- src/main/resources/de/tum/in/www1/jenkins/notifications/SendTestResultsNotificationPostBuildTask/config_de.properties
@@ -1,2 +1,0 @@
-Target=Benachrichtigungs URL
-Secret=Token
