--- src/main/resources/de/tum/in/www1/jenkins/notifications/SendTestResultsNotificationPostBuildTask/help-credentialsId.html
@@ -1,3 +1,0 @@
-<div>
-    The token authenticating the notification request against the specified URL
-</div>
