--- src/main/resources/de/tum/in/www1/jenkins/notifications/SendTestResultsNotificationPostBuildTask/help-credentialsId_de.html
@@ -1,3 +1,0 @@
-<div>
-    Das Token, welches die Benachrichtigung gegenüber der angegebenen URL authentifiziert.
-</div>
