--- src/main/resources/de/tum/in/www1/jenkins/notifications/SendTestResultsNotificationPostBuildTask/help-notificationUrl.html
@@ -1,3 +1,0 @@
-<div>
-    The URL to which the test results should get posted to.
-</div>
