--- src/main/resources/de/tum/in/www1/jenkins/notifications/SendTestResultsNotificationPostBuildTask/help-notificationUrl_de.html
@@ -1,3 +1,0 @@
-<div>
-    Die URL, an welche alle Test Ergebnisse gesendet werden sollen.
-</div>
