--- src/main/resources/index.jelly
@@ -1,4 +1,0 @@
-<?jelly escape-by-default='true'?>
-<div>
-    Send JUnit test result notifications to a REST API
-</div>
